Ported changes from 3.5.0 branch.
git-svn-id: http://svn.igniterealtime.org/svn/repos/openfire/trunk@10090 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/database/openfire_postgresql.sql
@@ -168,7 +168,7 @@ CREATE TABLE jiveSASLAuthorized (
 CREATE TABLE jiveSecurityAuditLog (
   msgID                 INTEGER         NOT NULL,
   username              VARCHAR(64)     NOT NULL,
-  entryStamp            INTEGER         NOT NULL,
+  entryStamp            BIGINT          NOT NULL,
   summary               VARCHAR(255)    NOT NULL,
   node                  VARCHAR(255)    NOT NULL,
   details               TEXT,
